Made five unique enemies with varying skill sets and enabled collisions.
I reused  the code for the first enemy to make more enemies. I also made
it so they each move at different rates, making the game more
challenging to the user. When the main character collides with an enemy,
the game ends. However, there is a glitch with the main character,
causing him to flicker out of a collision.